API/requirements.txt
flask
gspread
google-auth
python-dotenv
gunicorn
flask_cors

